refactor : 연관관계 매핑
